requirements.txt
click>=8.1.7
pyyaml>=6.0.2
requests>=2.32.3
beautifulsoup4>=4.12.3
python-whois>=0.9.4
dnspython>=2.6.1
colorama>=0.4.6
shodan>=1.31.0
aiohttp>=3.10.5
ratelimit>=2.2.1
rich>=13.9.1
prompt_toolkit>=3.0.47
weasyprint>=63.0
jinja2>=3.1.4
pyjwt>=2.9.0
injector>=0.22.0
RestrictedPython>=7.3
validators>=0.33.0